requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.0
websockets>=13.0
httpx>=0.27.0
playwright>=1.40.0
Pillow>=10.0.0